refactor(core): extract createArrowNavigationController (PLAN §3.1.h)
Pull arrow-key element navigation out of init() into a self-contained
controller (~96 lines moved):

- arrowNavigationElements + arrowNavigationActiveIndex signals
- arrowNavigator (the createArrowNavigator() spatial helper)
- clearArrowNavigation, selectAndFocusElement, openArrowNavigationMenu
- handleArrowNavigation (arrow keydown dispatch)
- handleArrowNavigationSelect (menu item click)
- arrowNavigationItems + arrowNavigationState memos

The controller takes (grab, phase, effectiveElement, isShiftMultiSelecting,
setKeyboardSelectedElement) and exposes its state + handlers. The
keyboardSelectedElement closure flag stays in init() (other handlers
also read/write it); the controller writes to it via a setter callback.

The contextmenu listener's `arrowNavigationElements().length > 0` check
now uses the controller's `state().isVisible` instead.

Drops 3 now-unused imports (getVisibleBoundsCenter, ArrowNavigationState
type, createArrowNavigator).

index.tsx: 3314 -> 3218 lines (-96). 21 keyboard-navigation specs green.
